+++ src/main/java/it/esercizio/banca_sella/dto/response/Balance.java
@@ -4,13 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Balance {
 
     private String date;
-    private Double balance;
-    private Double availableBalance;
+    private BigDecimal balance;
+    private BigDecimal availableBalance;
     private String currency;
 }
